make search functions local
add necessary search functions for all types of data, remove chained service dependencies, update page not found component

+++ src/app/notes/mynotes/mynotes.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { Note } from 'src/app/model/note';
 import { JwttokenService } from 'src/app/services/jwttoken.service';
 import { NotesService } from 'src/app/services/notes.service';
 
@@ -12,35 +13,53 @@ export class MynotesComponent implements OnInit {
 
   constructor(private notesService:NotesService, private _snackBar: MatSnackBar, private jwtTokenService: JwttokenService) { }
 
-  data: any;
+  data: Note[] = [];
+  displayedData: Note[] = [];
+  p: number = 1; // for pagination
+  searchValue: string = '';
 
   ngOnInit(): void {
     this.findAllByUsername();
   }
 
   public findAllByUsername(): any {
-    this.notesService.findAllByUsername(this.jwtTokenService.getUser().username).subscribe(value => { this.data = value; });
+    this.notesService.findAllByUsername(this.jwtTokenService.getUser().username).subscribe(value => { this.data = value; this.displayedData = this.data; });
   }
 
-  public deleteNote(id: string): any {
-    this.notesService.deleteNote(id).subscribe(data => { 
-      this._snackBar.open("Note deleted!", "", {duration: 2500});
-      setTimeout(()=>{this.findAllByUsername()}, 0);
-    });
-  }
-
-  public editNoteRedirect(id: string): any {
-    this.notesService.editNote(id);
-  }
-
-  public search(search: string) {
-    search = search.trim();
-    if (search != "") {
-      this.notesService.searchNotes(search, this.jwtTokenService.getUser().username).subscribe(result => { this.data = result; });
+  public deleteNote(id: string | undefined): any {
+    if (id) {
+      this.notesService.deleteNote(id).subscribe(data => { 
+        this._snackBar.open("Note deleted!", "", {duration: 2500});
+        setTimeout(()=>{this.findAllByUsername()}, 0);
+      });
     }
+  }
+
+  public editNoteRedirect(id: string | undefined): any {
+    id ? this.notesService.editNote(id) : null;
+  }
+
+  public searchNotes() {
+    let search = this.searchValue.trim().toLowerCase();
+    let arr = this.data;
+    this.p = 1;
+
+    if (search == '')
+      arr = this.data;
     else {
-      this.findAllByUsername();
+      arr = this.data.filter(obj => { 
+        return (obj.title.toLowerCase().includes(search) || ( obj.content ? obj.content.toLowerCase().includes(search) : null) ); 
+      });
+      this.p = 1;
     }
+
+    this.displayedData = arr;
+    // if (search != "") {
+    //   this.notesService.searchNotes(search, this.jwtTokenService.getUser().username).subscribe(result => { this.data = result; });
+    // }
+    // else {
+    //   this.findAllByUsername();
+    // }
   }
 
 }
